+++ src-test/org/etools/j1939_84/bus/j1939/packets/EngineHoursTimerTest.java
@@ -23,15 +23,13 @@ public class EngineHoursTimerTest {
 
     @Before
     public void setUp() {
-        byte[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, (byte) 0xFF, (byte) 0xFF,
-                (byte) 0xFF, (byte) 0xFF };
+        int[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0xFF, 0xFF };
         instance = new EngineHoursTimer(data);
     }
 
     @Test
     public void testEngineHoursTimerError() {
-        byte[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, (byte) 0xFE, (byte) 0xFE,
-                (byte) 0xFE, (byte) 0xFE };
+        int[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, 0xFE, 0xFE, 0xFE, 0xFE };
         EngineHoursTimer errorInstance = new EngineHoursTimer(data);
         assertEquals(ERROR, errorInstance.getEiAecdTimer2());
         String expected = "EI-AECD Number = 1: Timer 1 = 0 minutes; Timer 2 = errored";
@@ -53,17 +51,16 @@ public class EngineHoursTimerTest {
     @SuppressWarnings({ "SimplifiableAssertion", "EqualsWithItself" })
     @Test
     public void testEquals() {
-        byte[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, (byte) 0xFF, (byte) 0xFF,
-                (byte) 0xFF, (byte) 0xFF };
+        int[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0xFF, 0xFF };
         EngineHoursTimer expected = new EngineHoursTimer(data);
         assertTrue(instance.equals(expected));
-        byte[] data1 = { 0x38, 0x1A, 0x00, 0x00, 0x00, 0x0B, 0x00, 0x00, 0x00 };
+        int[] data1 = { 0x38, 0x1A, 0x00, 0x00, 0x00, 0x0B, 0x00, 0x00, 0x00 };
         EngineHoursTimer expected1 = new EngineHoursTimer(data1);
         assertFalse(instance.equals(expected1));
-        byte[] data2 = { 0x01, 0x1A, 0x00, 0x00, 0x00, 0x0B, 0x00, 0x00, 0x00 };
+        int[] data2 = { 0x01, 0x1A, 0x00, 0x00, 0x00, 0x0B, 0x00, 0x00, 0x00 };
         EngineHoursTimer expected2 = new EngineHoursTimer(data2);
         assertFalse(instance.equals(expected2));
-        byte[] data3 = { 0x01, 0x00, 0x00, 0x00, 0x00, 0x0B, 0x00, 0x00, 0x00 };
+        int[] data3 = { 0x01, 0x00, 0x00, 0x00, 0x00, 0x0B, 0x00, 0x00, 0x00 };
         EngineHoursTimer expected3 = new EngineHoursTimer(data3);
         assertFalse(instance.equals(expected3));
         assertFalse(instance.equals(new Object()));
@@ -88,8 +85,7 @@ public class EngineHoursTimerTest {
 
     @Test
     public void testHashCode() {
-        byte[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, (byte) 0xFF, (byte) 0xFF,
-                (byte) 0xFF, (byte) 0xFF };
+        int[] data = { 0x01, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0xFF, 0xFF };
         EngineHoursTimer expected = new EngineHoursTimer(data);
         assertEquals(expected.hashCode(), instance.hashCode());
     }
